Use activation handler for panel menu
